[PATRO23-7] move Keycloak settings to Domain
